feat: Add support for `min_primer_pair_hits` to `OffTargetDetector`  (#44)
Closes #43 

This PR introduces a new parameter to `OffTargetDetector`,
`min_primer_pair_hits`. This permits the user to relax the current
constraint that every primer pair have at least one hit in the
reference.

I have set the default value to 1, so default behavior remains unchanged
by this PR, and documented the use case where it may be appropriate to
change this value.