fix: Check against `HN` tag instead of `is_unmapped` property (#69)
Closes #68 

`bwa aln` produces inconsistent behavior when recording the mapped
status of multiply mapping reads. In most cases, such reads are flagged
as mapped with a mapq of 0. However, we have discovered some cases where
these reads are flagged as unmapped.

`prymer` currently relies on the unmapped flag to identify reads with a
hit count of zero. This has been yielding errors when attempting to
parse reads in the latter category. This information can instead be
obtained directly from the `HN` tag.

---

NB: Thanks to detective work by @tfenne 

> This happens because internally bwa concatenates all the sequences in
the FASTA file, so it can discover hits that extend across contig
boundaries. Unfortunately when it discovers it’s done this on the
mapping selected as the primary mapping, it just marks it as unmapped
instead of removing the mapping entirely (and selecting a different
primary mapping if there are secondary hits).

This PR additionally discards any hits that extend beyond the end of the
contig to which they are mapped.

---------

Co-authored-by: Tim Fennell <tfenne@tfenne.com>