Expose some bwa indel parameters through the off-target checker. (#95)
I did some playing with ntthal, and it appears to that the when
calculating Tms of duplexes, an indel closer than 3bp from the end of
the primer causes it to see the rest of the primer as non-annealed. In
some cases it can be more (depending on sequence, it may not see the
remainder as annealed after an indel 4-5bp in), but it seems consistent
that it needs _at least_ 3bp past the indel to anneal bases. Hence
setting `min_indel_to_end_distance: int = 3`.